third_party/blink/web_tests/fast/overflow/overflow-rtl-vertical-origin-expected.txt

This tests that the scroll origin for boxes with -webkit-writing-mode:vertical-rl; direction:rtl; overflow:scroll is correct.

PASS: test.scrollTop is correct.
